Struct atomic_write_file::AtomicWriteFile

pub struct AtomicWriteFile { /* private fields */ }

A file whose contents become visible to users only after the file is committed.

An AtomicWriteFile is a file that is assigned to a path, but whose contents won’t appear at that path until the file is committed. If AtomicWriteFile is used to open a file that already exists, the contents of the existing file will remain available until the AtomicWriteFile is committed. During that time, the AtomicWriteFile may be used to write new contents, but these new contents won’t be visible until after the file is committed.

Internally, AtomicWriteFile is implemented by initally opening a temporary file, and then renaming the temporary file to its final path on commit. See the module-level documentation for more details about the implementation.

An AtomicWriteFile is automatically discarded when it goes out of scope (when it gets dropped). Any error that occurs on drop is ignored. For this reason, if the file should not be committed, it is highly recommended that AtomicWriteFile is discarded explicitly using the discard() method, which allows callers to detect errors on cleanup. See committing or discarding changes below for more information.

Opening an AtomicWriteFile

There are two ways to obtain an AtomicWriteFile struct:

AtomicWriteFile::open()

OpenOptions::open()

The first method opens a file at the specified path with some default options. The second method using OpenOptions allows configuring how the file is opened.

Compatibility with std::fs::File

AtomicWriteFile implements the same methods and traits of std::fs::File, and aims to be as much compatible with File as possible. In fact, AtomicWriteFile can be dereferenced into a File struct: this means that you can use all methods provided by File directly on an AtomicWriteFile (just like you can use all of str methods on a String).

A reference to the wrapped File struct may also be explicitly obtained using as_file() and as_file_mut().

Committing or discarding changes

AtomicWriteFile provides two additional methods that are not provided by File: commit() and discard(). These methods can be called to save the new contents to the file path, or to destroy the new contents and leave the original file (if any) unchaged, respectively.

Changes are automatically discarded also when AtomicWriteFile is dropped. Therefore calling discard() is not mandatory, but it is highly recommended because the Drop implementation ignores all errors.

Cloning

Cloning a AtomicWriteFile is not possible, because this would result in ambiguity and race conditions when committing the file and its clones. It is however possible to clone the underlaying File struct using try_clone(). Writes to this cloned File however won’t be atomic after the AtomicWriteFile is committed.

Examples

Opening a file, writing new contents, and committing the changes:

use std::io::Write;
use atomic_write_file::AtomicWriteFile;

let mut file = AtomicWriteFile::open("foo.txt")?; // if "foo.txt" already exists, it is not
                                                  // initially truncated or deleted
writeln!(file, "hello")?; // "hello" is written to a temporary location; "foo.txt" (if it
                          // exists) keeps its old contents after this write

file.commit()?; // only now "foo.txt" gets swapped with the new contents ("hello")

Implementations

impl AtomicWriteFile

pub fn open<P: AsRef<Path>>(path: P) -> Result<AtomicWriteFile>

Opens an atomically-written file at path.

See OpenOptions for more details.

Examples

use atomic_write_file::AtomicWriteFile;
let file = AtomicWriteFile::open("foo.txt");

pub fn options() -> OpenOptions

Creates a new OpenOptions with default options.

This is equivalent to OpenOptions::new(), but allows for more readable code.

Examples

use atomic_write_file::AtomicWriteFile;
let file = AtomicWriteFile::options().read(true).open("foo.txt");

pub fn as_file(&self) -> &File

Returns a reference to the underlaying File struct.

The returned reference can be used to inspect or manipulate the contents and metadata of this AtomicWriteFile.

Examples

use std::os::fd::AsRawFd;
use atomic_write_file::AtomicWriteFile;

let file = AtomicWriteFile::open("foo.txt")?;
assert_eq!(file.as_raw_fd(), file.as_file().as_raw_fd());

pub fn as_file_mut(&mut self) -> &mut File

Returns a mutable reference to the underlaying File struct.

The returned reference can be used to inspect or manipulate the contents and metadata of this AtomicWriteFile.

Examples

use std::io::Write;
use atomic_write_file::AtomicWriteFile;

let mut file = AtomicWriteFile::open("foo.txt")?;
writeln!(file.as_file_mut(), "hello")?;

pub fn commit(self) -> Result<()>

Saves the contents of this file to its path.

After calling commit(), the AtomicWriteFile is consumed and can no longer be used. Clones of the underlaying File may still be used after calling commit(), although any write from that point onwards will no longer be atomic.

See the documentation for AtomicWriteFile and the module-level documentation for details about the internal implementation of commit(), as well as platform-specific details.

This method is automatically called when AtomicWriteFile is dropped, although in that case any error produced by commit() is ignored.

See also AtomicWriteFile::discard().

Examples

use std::io::Write;
use atomic_write_file::AtomicWriteFile;

let file = AtomicWriteFile::open("foo.txt")?;
writeln!(&file, "hello")?;
file.commit()?;

pub fn discard(self) -> Result<()>

Discard the contents of this file, and leave its path unchanged.

After calling discard(), the AtomicWriteFile is consumed and can no longer be used. Clones of the underlaying File may still be used after calling discard().

See also AtomicWriteFile::commit().

Examples

use std::io::Write;
use atomic_write_file::AtomicWriteFile;

let file = AtomicWriteFile::open("foo.txt")?;
writeln!(&file, "hello")?;
file.discard()?;